new hook to fetch data with a single ID - rename old fetch data hook and use old name for new hook
