+++ ccpp-jammming/src/components/ui/home/search_container/general_results/list_components/ArtistResultsBox.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ArtistResultItem from "./unit_components/ArtistResultItem";
-import useFetchContent from "../../../../../../hooks/useFetchContent";
+import useFetchSearchResults from "../../../../../../hooks/useFetchSearchResults";
 
 function ArtistResultsBox({ searchArtistResults, searchAlbumResults, searchTrackResults, onArtistClick, onAlbumClick, accessToken }) {
-    const { fetchedArtistsArray } = useFetchContent({ searchArtistResults, accessToken })
+    const { fetchedArtistsArray } = useFetchSearchResults({ searchArtistResults, accessToken })
 
     return (
         <>
@@ -12,10 +12,11 @@ function ArtistResultsBox({ searchArtistResults, searchAlbumResults, searchTrack
                 <ArtistResultItem 
                     artistContent={artist}
                     onArtistClick={onArtistClick}
+                    accessToken={accessToken}
                     key={artist.artistUri}/>
             ))}
         </>
     );
 }
 
-export default ArtistResultsBox;
+export default ArtistResultsBox;
